Change type of NetworkBehaviour's "size" field to "short", to save a few bytes when transferring it across the network

+++ src/main/java/com/ezardlabs/dethsquare/multiplayer/NetworkAnimator.java
@@ -26,7 +26,7 @@ public class NetworkAnimator extends NetworkBehaviour {
 	}
 
 	@Override
-	public int getSize() {
+	public short getSize() {
 		return 8;
 	}
 }
